+++ main.go
@@ -20,6 +20,7 @@ import (
 	"crypto/tls"
 	"flag"
 	"fmt"
+	"net/url"
 	"os"
 	"time"
 
@@ -63,6 +64,9 @@ func init() {
 }
 
 func main() {
+	var grafanaURL string
+	var err error
+
 	flag.StringVar(&conf.MetricsAddr, "metrics-bind-address", ":8080",
 		"The address the metric endpoint binds to.")
 	flag.StringVar(&conf.ProbeAddr, "health-probe-bind-address", ":8081",
@@ -76,6 +80,8 @@ func main() {
 		"If set, HTTP/2 will be enabled for the metrics and webhook servers")
 	flag.StringVar(&conf.OperatorNamespace, "operator-namespace", "",
 		"The namespace where the observability-operator is running.")
+	flag.StringVar(&grafanaURL, "grafana-url", "http://grafana.monitoring.svc.cluster.local",
+		"grafana URL")
 
 	// Management cluster configuration flags.
 	flag.StringVar(&conf.ManagementCluster.BaseDomain, "management-cluster-base-domain", "",
@@ -110,6 +116,8 @@ func main() {
 		"The version of Prometheus Agents to deploy.")
 	flag.DurationVar(&conf.Monitoring.WALTruncateFrequency, "monitoring-wal-truncate-frequency", 2*time.Hour,
 		"Configures how frequently the Write-Ahead Log (WAL) truncates segments.")
+	flag.StringVar(&conf.Monitoring.MetricsQueryURL, "monitoring-metrics-query-url", "http://mimir-gateway.mimir.svc/prometheus",
+		"URL to query for cluster metrics")
 	opts := zap.Options{
 		Development: false,
 	}
@@ -117,10 +125,16 @@ func main() {
 	opts.BindFlags(flag.CommandLine)
 	flag.Parse()
 
+	// parse grafana URL
+	conf.GrafanaURL, err = url.Parse(grafanaURL)
+	if err != nil {
+		panic(fmt.Sprintf("failed to parse grafana url: %v", err))
+	}
+
 	ctrl.SetLogger(zap.New(zap.UseFlagOptions(&opts)))
 
 	// Load environment variables.
-	_, err := env.UnmarshalFromEnviron(&conf.Environment)
+	_, err = env.UnmarshalFromEnviron(&conf.Environment)
 	if err != nil {
 		setupLog.Error(err, "failed to unmarshal environment variables")
 		os.Exit(1)
@@ -185,7 +199,7 @@ func main() {
 	}
 
 	// Setup controller for the GrafanaOrganization resource.
-	err = controller.SetupGrafanaOrganizationReconciler(mgr, conf.Environment)
+	err = controller.SetupGrafanaOrganizationReconciler(mgr, conf)
 	if err != nil {
 		setupLog.Error(err, "unable to setup controller", "controller", "GrafanaOrganizationReconciler")
 		os.Exit(1)