+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- command line args to configure mimir and grafana URLs
+
 ## [0.10.2] - 2024-12-17
 
 ### Added